Move qubit management transformers from Cirq-FT to Cirq-core
